+++ tests/networktests.cpp
@@ -1,4 +1,6 @@
 #include <gtest/gtest.h>
+#include "network/ipv4address.h"
+#include "network/macaddress.h"
 #include "network/networkmonitor.h"
 #include "network/webclient.h"
 #include "system/environment.h"
@@ -52,4 +54,62 @@ TEST_F(NetworkTest, FetchJsonString1)
 {
     std::string s{ m_webClient->fetchJson("https://api.github.com/repos/nickvisionapps/denaro/tags") };
     ASSERT_TRUE(!s.empty());
+}
+
+TEST_F(NetworkTest, IPv4Address1)
+{
+    IPv4Address ip{ 192, 168, 1, 1 };
+    ASSERT_EQ(ip.getFirst(), 192);
+    ASSERT_EQ(ip.getSecond(), 168);
+    ASSERT_EQ(ip.getThird(), 1);
+    ASSERT_EQ(ip.getFourth(), 1);
+    ASSERT_EQ(ip.str(), "192.168.1.1");
+}
+
+TEST_F(NetworkTest, IPv4Address2)
+{
+    std::optional<IPv4Address> ip{ IPv4Address::parse("76.98.1.193") };
+    ASSERT_TRUE(ip.has_value());
+    ASSERT_EQ(ip->getFirst(), 76);
+    ASSERT_EQ(ip->getSecond(), 98);
+    ASSERT_EQ(ip->getThird(), 1);
+    ASSERT_EQ(ip->getFourth(), 193);
+    ASSERT_EQ(ip->str(), "76.98.1.193");
+}
+
+TEST_F(NetworkTest, IPv4Address3)
+{
+    std::optional<IPv4Address> ip{ IPv4Address::parse("a.b.c.d") };
+    ASSERT_FALSE(ip.has_value());
+}
+
+TEST_F(NetworkTest, MacAddress1)
+{
+    MacAddress mac{ 0x00, 0x1A, 0x2B, 0x3C, 0x4D, 0x5E };
+    ASSERT_EQ(mac.getFirst(), 0x00);
+    ASSERT_EQ(mac.getSecond(), 0x1A);
+    ASSERT_EQ(mac.getThird(), 0x2B);
+    ASSERT_EQ(mac.getFourth(), 0x3C);
+    ASSERT_EQ(mac.getFifth(), 0x4D);
+    ASSERT_EQ(mac.getSixth(), 0x5E);
+    ASSERT_EQ(mac.str(), "00:1A:2B:3C:4D:5E");
+}
+
+TEST_F(NetworkTest, MacAddress2)
+{
+    std::optional<MacAddress> mac{ MacAddress::parse("00:1A:2B:3C:4D:5E") };
+    ASSERT_TRUE(mac.has_value());
+    ASSERT_EQ(mac->getFirst(), 0x00);
+    ASSERT_EQ(mac->getSecond(), 0x1A);
+    ASSERT_EQ(mac->getThird(), 0x2B);
+    ASSERT_EQ(mac->getFourth(), 0x3C);
+    ASSERT_EQ(mac->getFifth(), 0x4D);
+    ASSERT_EQ(mac->getSixth(), 0x5E);
+    ASSERT_EQ(mac->str(), "00:1A:2B:3C:4D:5E");
+}
+
+TEST_F(NetworkTest, MacAddress3)
+{
+    std::optional<MacAddress> mac{ MacAddress::parse("00:1A:2B:3C:4D") };
+    ASSERT_FALSE(mac.has_value());
 }