+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 #### Events
 - Added `Nickvision::Events::Event::count()` method
 - Added `Nickvision::Events::Event::operator bool()` method
+#### Network
+- Added `Nickvision::Network::IPv4Address` class
+- Added `Nickvision::Network::MacAddress` class
 ### Fixes
 None